Treat a SIM power cycle as first boot.
TL;DR - forget the previous subId used only for optimization
purposes if we get a SIM_READY notification.

We moved some code that poked the modem from a SIM_READY handler
to a onSubscriptionChanged handler and also protected it
with logic to only fire if the subId changed.  This caused
us to not poke the modem on SIM power cycle.  Some modems
do a SIM power cycle on airplane mode.  The result was
on these modems we weren't causing the modem to register with
the network after airplane mode.

This was often masked because if we were setup to try a
data call these devices were configured to try the data call
regardless of data registration and the act of trying a data call
would cause the modem to register with the network.

If during airplane mode you had turned on wifi, telephony would
stop trying to setup a data call (so we don't poke the network
all the time when on a better wifi connection) and the restoration of
that data request has been blocked until data-registration in the
new multisim code.

bug:19194287
Change-Id: I21b76b7c62a6161d8422b11a831c3747591c56f3

Conflicts:
	src/java/com/android/internal/telephony/ServiceStateTracker.java
	src/java/com/android/internal/telephony/gsm/GsmServiceStateTracker.java